+++ src/Ninja/Program.cs
@@ -3,13 +3,20 @@
 
 using System.CommandLine;
 using Microsoft.NugetNinja;
-using Microsoft.NugetNinja.Core;
 using Microsoft.NugetNinja.Framework;
+using Microsoft.NugetNinja.UselessPackageReferencePlugin;
 
 var rootCommand = new RootCommand(@"Nuget Ninja, a tool for detecting dependencies of .NET projects.")
     .AddGlobalOptions();
 
-foreach (var subCommandInstance in StartUp.GetCommandHandlers())
+var handlers = new CommandHandler[]
+{ 
+    new PackageReferenceHandler<StartUp>(),
+    new ProjectReferenceHandler<StartUp>(),
+    new PackageUpgradeHandler<StartUp>()
+};
+
+foreach (var subCommandInstance in handlers)
 {
     rootCommand.Add(subCommandInstance.BuildAsCommand());
 }
